training/requirements.txt
stable-baselines3>=2.3.0
gymnasium>=0.29.0
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.16.0
pyyaml>=6.0
numpy>=1.24.0
